Mean Functions

Mean Functions

The package provides functions to compute means of different kinds.

StatsBase.geomean — Function.

geomean(a)

Return the geometric mean of a real-valued array.

source

StatsBase.harmmean — Function.

harmmean(a)

Return the harmonic mean of a real-valued array.

source

StatsBase.genmean — Function.

genmean(a, p)

Return the generalized/power mean with exponent p of a real-valued array, i.e. $\left( \frac{1}{n} \sum_{i=1}^n a_i^p \right)^{\frac{1}{p}}$, where n = length(a). It is taken to be the geometric mean when p == 0.

source

The mean and mean! functions are also extended to accept a weight vector of type AbstractWeights to compute weighted mean.

Base.mean — Method.

mean(A::AbstractArray, w::AbstractWeights[, dim::Int])

Compute the weighted mean of array A with weight vector w (of type AbstractWeights). If dim is provided, compute the weighted mean along dimension dim.

Examples

w = rand(n)
mean(x, weights(w))

source

Base.mean! — Method.

mean(R::AbstractArray, , A::AbstractArray, w::AbstractWeights[, dim::Int])

Compute the weighted mean of array A with weight vector w (of type AbstractWeights) along dimension dim, and write results to R.

source