Abstraction for Statistical Models

Abstraction for Statistical Models

This package defines an abstract type StatisticalModel, and an abstract subtype RegressionModel.

Particularly, instances of StatisticalModel implement the following methods.

StatsBase.adjr2 — Function.

adjr2(obj::StatisticalModel)
adjr²(obj::StatisticalModel)

Adjusted coefficient of determination (adjusted R-squared).

For linear models, the adjusted R² is defined as $1 - (1 - (1-R^2)(n-1)/(n-p))$, with $R^2$ the coefficient of determination, $n$ the number of observations, and $p$ the number of coefficients (including the intercept). This definition is generally known as the Wherry Formula I.

source

adjr2(obj::StatisticalModel, variant::Symbol)
adjr²(obj::StatisticalModel, variant::Symbol)

Adjusted pseudo-coefficient of determination (adjusted pseudo R-squared).

For nonlinear models, one of the several pseudo R² definitions must be chosen via variant. The only currently supported variant is :MacFadden, defined as $1 - (\log (L) - k)/\log (L0)$. In this formula, $L$ is the likelihood of the model, $L0$ that of the null model (the model including only the intercept). These two quantities are taken to be minus half deviance of the corresponding models. $k$ is the number of consumed degrees of freedom of the model (as returned by dof).

source

StatsBase.aic — Function.

aic(obj::StatisticalModel)

Akaike's Information Criterion, defined as $-2 \log L + 2k$, with $L$ the likelihood of the model, and k its number of consumed degrees of freedom (as returned by dof).

source

StatsBase.aicc — Function.

aicc(obj::StatisticalModel)

Corrected Akaike's Information Criterion for small sample sizes (Hurvich and Tsai 1989), defined as $-2 \log L + 2k + 2k(k-1)/(n-k-1)$, with $L$ the likelihood of the model, $k$ its number of consumed degrees of freedom (as returned by dof), and $n$ the number of observations (as returned by nobs).

source

StatsBase.bic — Function.

bic(obj::StatisticalModel)

Bayesian Information Criterion, defined as $-2 \log L + k \log n$, with $L$ the likelihood of the model, $k$ its number of consumed degrees of freedom (as returned by dof), and $n$ the number of observations (as returned by nobs).

source

StatsBase.coef — Function.

coef(obj::StatisticalModel)

Return the coefficients of the model.

source

StatsBase.coefnames — Function.

coefnames(obj::StatisticalModel)

Return the names of the coefficients.

source

StatsBase.coeftable — Function.

coeftable(obj::StatisticalModel; level::Real=0.95)

Return a table of class CoefTable with coefficients and related statistics. level determines the level for confidence intervals (by default, 95%).

source

StatsBase.confint — Function.

confint(obj::StatisticalModel; level::Real=0.95)

Compute confidence intervals for coefficients, with confidence level level (by default 95%).

source

StatsBase.deviance — Function.

deviance(obj::StatisticalModel)

Return the deviance of the model relative to a reference, which is usually when applicable the saturated model. It is equal, up to a constant, to $-2 \log L$, with $L$ the likelihood of the model.

source

StatsBase.dof — Function.

dof(obj::StatisticalModel)

Return the number of degrees of freedom consumed in the model, including when applicable the intercept and the distribution's dispersion parameter.

source

StatsBase.fit — Function.

fit(Histogram, data[, weight][, edges]; closed=:left, nbins)

Fit a histogram to data.

Arguments

data: either a vector (for a 1-dimensional histogram), or a tuple of vectors of equal length (for an n-dimensional histogram).

weight: an optional AbstractWeights (of the same length as the data vectors), denoting the weight each observation contributes to the bin. If no weight vector is supplied, each observation has weight 1.

edges: a vector (typically an AbstractRange object), or tuple of vectors, that gives the edges of the bins along each dimension. If no edges are provided, these are determined from the data.

Keyword arguments

closed: if :left (the default), the bin intervals are left-closed [a,b); if :right, intervals are right-closed (a,b].

nbins: if no edges argument is supplied, the approximate number of bins to use along each dimension (can be either a single integer, or a tuple of integers).

Examples

# Univariate
h = fit(Histogram, rand(100))
h = fit(Histogram, rand(100), 0:0.1:1.0)
h = fit(Histogram, rand(100), nbins=10)
h = fit(Histogram, rand(100), weights(rand(100)), 0:0.1:1.0)
h = fit(Histogram, [20], 0:20:100)
h = fit(Histogram, [20], 0:20:100, closed=:right)

# Multivariate
h = fit(Histogram, (rand(100),rand(100)))
h = fit(Histogram, (rand(100),rand(100)),nbins=10)

source

Fit a statistical model.

source

fit(ZScoreTransform, X; center=true, scale=true)

Fit standardization parameters to X and return a ZScoreTransform transformation object.

Arguments

data: matrix of samples to fit transformation parameters.

Keyword arguments

center: if true (the default) center data so that its mean is zero.

scale: if true (the default) scale the data so that its variance is equal to one.

Examples

julia> using StatsBase

julia> X = [0.0 -0.5 0.5; 0.0 1.0 2.0]
2×3 Array{Float64,2}:
 0.0  -0.5  0.5
 0.0   1.0  2.0

julia> dt = fit(ZScoreTransform, X)
ZScoreTransform{Float64}(2, [0.0, 1.0], [0.5, 1.0])

julia> StatsBase.transform(dt, X)
2×3 Array{Float64,2}:
  0.0  -1.0  1.0
 -1.0   0.0  1.0

source

fit(UnitRangeTransform, X; center=true, scale=true)

Fit a scaling parameters to X and return transformation description.

Arguments

data: matrix of samples to fit transformation parameters.

Keyword arguments

center: if true (the default) centere data around zero.

scale: if true (the default) perform variance scaling.

Examples

julia> using StatsBase

julia> X = [0.0 -0.5 0.5; 0.0 1.0 2.0]
2×3 Array{Float64,2}:
 0.0  -0.5  0.5
 0.0   1.0  2.0

julia> dt = fit(UnitRangeTransform, X)
UnitRangeTransform{Float64}(2, true, [-0.5, 0.0], [1.0, 0.5])

julia> StatsBase.transform(dt, X)
2×3 Array{Float64,2}:
 0.5  0.0  1.0
 0.0  0.5  1.0

source

StatsBase.fit! — Function.

Fit a statistical model in-place.

source

StatsBase.informationmatrix — Function.

informationmatrix(model::StatisticalModel; expected::Bool = true)

Return the information matrix. By default the Fisher information matrix is returned, while the observed information matrix can be requested with expected = false.

source

StatsBase.isfitted — Function.

isfitted(obj::StatisticalModel)

Indicate whether the model has been fitted.

source

StatsBase.islinear — Function.

islinear(obj::StatisticalModel)

Indicate whether the model is linear.

source

StatsBase.loglikelihood — Function.

loglikelihood(obj::StatisticalModel)

Return the log-likelihood of the model.

source

StatsBase.mss — Function.

mss(obj::StatisticalModel)

Return the model sum of squares.

source

StatsBase.nobs — Function.

nobs(obj::StatisticalModel)

Return the number of independent observations on which the model was fitted. Be careful when using this information, as the definition of an independent observation may vary depending on the model, on the format used to pass the data, on the sampling plan (if specified), etc.

source

StatsBase.nulldeviance — Function.

nulldeviance(obj::StatisticalModel)

Return the deviance of the null model, that is the one including only the intercept.

source

StatsBase.r2 — Function.

r2(obj::StatisticalModel)
r²(obj::StatisticalModel)

Coefficient of determination (R-squared).

For a linear model, the R² is defined as $ESS/TSS$, with $ESS$ the explained sum of squares and $TSS$ the total sum of squares.

source

r2(obj::StatisticalModel, variant::Symbol)
r²(obj::StatisticalModel, variant::Symbol)

Pseudo-coefficient of determination (pseudo R-squared).

For nonlinear models, one of several pseudo R² definitions must be chosen via variant. Supported variants are:

:MacFadden (a.k.a. likelihood ratio index), defined as $1 - \log (L)/\log (L_0)$;

:CoxSnell, defined as $1 - (L_0/L)^{2/n}$;

:Nagelkerke, defined as $(1 - (L_0/L)^{2/n})/(1 - L_0^{2/n})$.

In the above formulas, $L$ is the likelihood of the model, $L_0$ is the likelihood of the null model (the model with only an intercept), $n$ is the number of observations, $y_i$ are the responses, $\hat{y}_i$ are fitted values and $\bar{y}$ is the average response.

Cox and Snell's R² should match the classical R² for linear models.

source

StatsBase.rss — Function.

rss(obj::StatisticalModel)

Return the residual sum of squares.

source

StatsBase.score — Function.

score(obj::StatisticalModel)

Return the score of the statistical model. The score is the gradient of the log-likelihood with respect to the coefficients.

source

StatsBase.stderror — Function.

stderror(obj::StatisticalModel)

Return the standard errors for the coefficients of the model.

source

StatsBase.vcov — Function.

vcov(obj::StatisticalModel)

Return the variance-covariance matrix for the coefficients of the model.

source

StatsBase.weights — Method.

weights(obj::StatisticalModel)

Return the weights used in the model.

source

RegressionModel extends StatisticalModel by implementing the following additional methods.

StatsBase.dof_residual — Function.

dof_residual(obj::RegressionModel)

Return the residual degrees of freedom of the model.

source

StatsBase.fitted — Function.

fitted(obj::RegressionModel)

Return the fitted values of the model.

source

StatsBase.leverage — Function.

leverage(obj::RegressionModel)

Return the diagonal of the projection matrix.

source

StatsBase.meanresponse — Function.

meanresponse(obj::RegressionModel)

Return the mean of the response.

source

StatsBase.modelmatrix — Function.

modelmatrix(obj::RegressionModel)

Return the model matrix (a.k.a. the design matrix).

source

StatsBase.response — Function.

response(obj::RegressionModel)

Return the model response (a.k.a. the dependent variable).

source

StatsBase.predict — Function.

predict(obj::RegressionModel, [newX])

Form the predicted response of model obj. An object with new covariate values newX can be supplied, which should have the same type and structure as that used to fit obj; e.g. for a GLM it would generally be a DataFrame with the same variable names as the original predictors.

source

StatsBase.predict! — Function.

predict!

In-place version of predict.

source

StatsBase.residuals — Function.

residuals(obj::RegressionModel)

Return the residuals of the model.

source